requirements.txt
phidata
python-dotenv
yfinance 
packaging
duckduckgo-search
fastapi
uvicorn
groq
